Accept action param with dashes
The default struts config disallows param names made up of words
separated by dashes, which we can't avoid because of the use of the
conventions plugin.

It's not clear to me exactly why bulk-suspend and bulk-delete aren't
affected by this; it probably has to do with the extra dialog forms
that pop up when clicking those buttons.